Update README.md
RESEARCH METHODOLOGY 
Given that measuring cracks manually takes a long time for large-scale constructions(e.g. high-rise buildings and bridges), Researchers have created image recognition models that provide a quicker and more precise technique to measure cracks on concrete surfaces. After installation, the model is used to spot and examine surface fractures in a variety of building materials, including brick walls with mortar covering them, doors, walls, columns, and slabs of concrete.
MATLAB operation
A high-performance language for technical computing is called MATLAB. It incorporates programming environment, calculation, and visualization. Additionally, MATLAB is a cutting-edge programming language environment that supports object-oriented programming, has sophisticated data structures, and has built-in editing and debugging tools. These elements make MATLAB a superior research and teaching tool.
1. Adjusting contrast
The gamma correction method alters an image's intensity values by increasing each intensity value to a particular power, or gamma value. The gamma value can be used to increase or decrease the contrast in an image by controlling the image's intensity distribution's form.
2. RGB to Grayscale
The "rgb2gray" command in MATLAB is used to convert RGB images to grayscale. By computing a weighted average of the red, green, and blue color channels, this function transforms an RGB image from color to grayscale. The red channel is given a weight of 0.2989, the green channel is given a weight of 0.5870, and the blue channel is given a weight of 0.1140. The relative sensitivity of the human visual system to various hues is the basis for these weights.
3. Grayscale Image to a Binary Image
The "im2bw" command in MATLAB is used to transform a grayscale picture into a binary image. By thresholding the image's intensity values, this function converts an image to binary. Pixels are set to 1 (white) if their intensity values are more than a certain threshold, and 0 if they are lower (black).
4. Flood-Fill Operations
A flood-fill process is carried out on binary and grayscale pictures by the imfill function. The removal of pointless artifacts from photos can benefit from this technique. This function fills in any gaps in the object by substituting foreground pixels for the background pixels within. Starting with an edge pixel, the function scans the image and fills any gaps that are connected to the backdrop.
Imfill converts linked background pixels (0s) to foreground pixels (1s) in binary images before terminating at object boundaries.
Imfill raises the intensity ratings of dark regions surrounded by lighter regions in grayscale photos to the same level as neighboring pixels. Regional minima that are not connected to the image border are essentially removed by imfill..(MathWorks - Makers of MATLAB and Simulink - MATLAB & Simulink, n.d.)
5. Morphological filter:
A kind of image processing filter known as a morphological filter works with binary or grayscale pictures. A morphological filter is used to alter the form or organization of objects in a picture or to draw out structural data from an image. This is done by performing fundamental set operations on the binary or grayscale image, such as dilation, erosion, opening, closure, etc. The functions in the Image Processing Toolbox in MATLAB, such as "imdilate," "imerode," "imopen," and "imclose," can be used to apply morphological filters. On a binary or grayscale image, these functions execute dilation, erosion, opening, and closure operations, respectively.
6. Median filter:
The intensity values of a pixel are changed by a median filter, a type of image processing filter, to the median intensity values of the pixels around it. Salt-and-pepper noise, a sort of impulsive noise in which some pixels are arbitrarily set to either the maximum or minimum intensity value, is frequently eliminated with this filter. The median filter operates by moving a window of a specific size over the image and determining the median of the intensities inside the window at each place. The middle pixel of the window is then given the median intensity value. All of the pixels in the image go through this process again, producing a filtered image with less noise. A median filter can be used in MATLAB.